+++ CHANGELOG.md
@@ -23,8 +23,9 @@ Extends support of the SDK to OneView REST API version 1200 (OneView v5.00).
   Refer example directory for more examples.
 
 #### Features supported with the current release
-- Network set
 - Interconnect
+- Network set
+- Storage system
 
 # TODO: Move the below resources to the above list once the resource module is updated to use API versions 1000 and 1200
 - Connection template